remove unusable data
We got rid of some of the user data because it was invalid or was irreparably corrupted.
We chose not to use 499b.csv because we weren't sure what the data was or if we just append it or not, and 499.csv already seemed to have a good amount of data anyway.
507.csv and 520.csv seemed to contain the old device_uplink
While 507b.csv did contain the data we were looking for, most of it's start and end values completley corrupted, showing only "##########"
Same for 536b.csv and 634.csv
530b.csv, 545.csv, 656.csv, 669.csv, 676.csv, 512b.csv, 635.csv, and 655b.csv are omitted for having too few entries in the data.
Our reasoning is that it might skew our models. Also, some of these files did not even have enough rows to account for even one day, so the data would be useless anyway.
230.csv and 230b.csv contain identical information; same for 514.csv and 514b.csv; and for 574.csv and 574b.csv. In order to not overrepresent this data, we chose to remove the b counterparts to each dataset.